requirements.txt
numpy==1.23.5
numpy==1.24.3
Pillow==9.5.0
Pillow==10.0.1
Requests==2.31.0
streamlit==1.26.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
